requirements.txt
numpy~=1.24.4
pandas~=2.0.3
torch~=2.4.1
pillow~=10.4.0
scikit-learn~=1.3.2
streamlit~=1.40.1
requests~=2.32.3
datasets~=3.1.0
pytesseract~=0.3.13
pdfplumber~=0.11.5
plotly~=6.0.0
matplotlib~=3.7.5
fitz~=0.0.1.dev2
transformers~=4.46.3
python-dotenv~=1.0.1